moved over the stepdetailactivty to use the videofragment to reduce code. cleaned up a lot of code. fixed instancestates on the stepactivity. changed up some layouts. added an exoplayer controller layout.
