+++ app/src/main/java/com/charlesrowland/ragingclaw/bakingapp/utils/AllMyConstants.java
@@ -10,9 +10,11 @@ public class AllMyConstants {
     public final static String RECIPE_NAME_EXTRA = "recipe_name_extra";
     public final static String RECIPE_INTENT_EXTRA = "recipe_intent_extra";
     public final static String RECIPE_INTENT_JSON_EXTRA = "recipe_intent_json_extra";
+    public final static String CURRENT_RECIPE = "current_recipe";
     public final static String INGREDIENTS_RECYCLER_STATE = "ingredients_recycler_state";
     public static final String ARG_ITEM_ID = "item_id";
     public final static String STEPS_ARRAYLIST_STATE = "steps_arraylist_state";
+    public final static String STEP_TITLE = "step_title";
     public final static String STEP_NUMBER = "step_number";
     public static final String STEP_URI =  "step_uri";
     public static final String STEP_VIDEO_POSITION =  "step_video_position";
