+++ app/src/main/res/layout/step_detail.xml
@@ -4,9 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     style="@style/CardView.Light"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="@dimen/margins_8">
-
+    android:layout_height="wrap_content">
 
     <include
         android:id="@+id/video_layout_container"
